+++ src/osgEarthDrivers/engine_osgterrain/CustomTile.cpp
@@ -833,7 +833,7 @@ CustomTile::servicePendingElevationRequests( const MapFrame& mapf, int stamp, bo
         // otherwise, see if it is legal yet to start a new request:
         else if ( readyForNewElevation() )
         {
-            if ( _elevationLOD + 1 == (int)_key.getLevelOfDetail() )
+            if ( _elevationLOD + 1 == _key.getLevelOfDetail() )
             {
                 _elevRequest->setStamp( stamp );
                 _elevRequest->setProgressCallback( new ProgressCallback() );
